+++ src/autocomplete/module.ts
@@ -4,7 +4,6 @@ import { FormsModule } from '@angular/forms';
 import { ThyAutocompleteTriggerDirective } from './autocomplete.trigger.directive';
 import { ThyInputModule } from '../input/module';
 import { ThyAutocompleteComponent } from './autocomplete.component';
-import { ThyAutoOptionComponent } from './option.component';
 import { ThyLabelModule } from '../label/label.module';
 import { OverlayModule } from '@angular/cdk/overlay';
 import { ThyLoadingModule } from '../loading';
@@ -13,6 +12,7 @@ import { ThyIconModule } from '../icon';
 import { ThyEmptyModule } from '../empty/empty.module';
 import { ThyAutocompleteContainerComponent } from './overlay/autocomplete-container.component';
 import { THY_AUTOCOMPLETE_DEFAULT_CONFIG_PROVIDER } from './overlay/autocomplete.config';
+import { ThyOptionModule } from '../core/option/module';
 import { PortalModule } from '@angular/cdk/portal';
 
 @NgModule({
@@ -26,21 +26,12 @@ import { PortalModule } from '@angular/cdk/portal';
         ThyLoadingModule,
         ThyDirectiveModule,
         ThyIconModule,
-        ThyEmptyModule
+        ThyEmptyModule,
+        ThyOptionModule
     ],
-    declarations: [
-        ThyAutocompleteTriggerDirective,
-        ThyAutocompleteComponent,
-        ThyAutoOptionComponent,
-        ThyAutocompleteContainerComponent
-    ],
+    declarations: [ThyAutocompleteTriggerDirective, ThyAutocompleteComponent, ThyAutocompleteContainerComponent],
     entryComponents: [ThyAutocompleteContainerComponent],
-    exports: [
-        ThyAutocompleteTriggerDirective,
-        ThyAutocompleteComponent,
-        ThyAutoOptionComponent,
-        ThyAutocompleteContainerComponent
-    ],
+    exports: [ThyAutocompleteTriggerDirective, ThyAutocompleteComponent, ThyAutocompleteContainerComponent],
     providers: [THY_AUTOCOMPLETE_DEFAULT_CONFIG_PROVIDER]
 })
 export class ThyAutocompleteModule {}