+++ src/core/select/select-option/select-option-base.ts
@@ -1,5 +1,0 @@
-export abstract class SelectOptionBase {
-    thyLabelText: string;
-    thyRawValue: any;
-    thyValue: any;
-}